requirements.txt
aiogram==3.21.0
python-dotenv==1.1.1
aiofiles==23.2.1